thêm cn quản lí địa chỉ, cập nhật chọn địa chỉ nhanh khi thanh toán

+++ database/migrations/2025_07_12_103915_add_user_id_to_shipping_addresses_table.php
@@ -9,13 +9,14 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
+public function up(): void
+{
+    if (!Schema::hasColumn('shipping_addresses', 'user_id')) {
         Schema::table('shipping_addresses', function (Blueprint $table) {
             $table->foreignId('user_id')->after('id')->constrained()->onDelete('cascade');
         });
     }
-
+}
     /**
      * Reverse the migrations.
      */
